Converted facial model to tensorflow
