+++ models/face/checkpoints/best_epochs.txt
@@ -1,0 +1,1 @@
+Binary: Epoch 9
