+++ models/face/checkpoints/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "binary.ckpt"
+all_model_checkpoint_paths: "binary.ckpt"
