+++ requirements.txt
@@ -2,6 +2,5 @@ moviepy==1.0.3
 numpy==1.26.1
 opencv-python==4.8.1.78
 pytest==7.4.3
-sklearn==0.0.post11
-torch==1.9.1
-torchvision==0.10.1
+scikit-learn==1.3.2
+tensorflow==2.15.0
